Relocating original event stream service.
