+++ test/factories/eventStreamServiceFactory.spec.ts
@@ -5,7 +5,7 @@ import { ErrorFactory } from '../../src/errors/errorFactory';
 import { EventStreamServiceFactoryImpl } from '../../src/factories/eventStreamServiceFactory';
 import { DeviceType } from '../../src/model';
 import { PlatformContainer } from '../../src/primitives/platformContainer';
-import { EventStreamServiceImpl } from '../../src/services/husqvarna/eventStreamService';
+import { AutomowerEventStreamService } from '../../src/services/husqvarna/automower/automowerEventStreamService';
 import { GardenaEventStreamService } from '../../src/services/husqvarna/gardena/gardenaEventStreamService';
 
 describe('EventStreamServiceFactoryImpl', () => {
@@ -27,22 +27,22 @@ describe('EventStreamServiceFactoryImpl', () => {
         target = new EventStreamServiceFactoryImpl(config, errorFactory.object());
     });
 
-    it('resolve the event stream service by default', () => {
-        const expected = new Mock<EventStreamServiceImpl>();
+    it('resolve the automower event stream service by default', () => {
+        const expected = new Mock<AutomowerEventStreamService>();
 
         config.device_type = undefined;
-        container.setup(o => o.resolve(EventStreamServiceImpl)).returns(expected.object());
+        container.setup(o => o.resolve(AutomowerEventStreamService)).returns(expected.object());
 
         const result = target.create(container.object());
 
         expect(result).toBe(expected.object());
     });
 
-    it('resolve the event stream service by automower service', () => {
-        const expected = new Mock<EventStreamServiceImpl>();
+    it('resolve the automower event stream service by automower service', () => {
+        const expected = new Mock<AutomowerEventStreamService>();
 
         config.device_type = DeviceType.AUTOMOWER;
-        container.setup(o => o.resolve(EventStreamServiceImpl)).returns(expected.object());
+        container.setup(o => o.resolve(AutomowerEventStreamService)).returns(expected.object());
 
         const result = target.create(container.object());
 
@@ -50,7 +50,7 @@ describe('EventStreamServiceFactoryImpl', () => {
     });
 
     it('resolve the gardena event stream service', () => {
-        const expected = new Mock<EventStreamServiceImpl>();
+        const expected = new Mock<GardenaEventStreamService>();
 
         config.device_type = DeviceType.GARDENA;
         container.setup(o => o.resolve(GardenaEventStreamService)).returns(expected.object());
