+++ test/services/husqvarna/automower/automowerEventStreamServiceSpy.ts
@@ -1,7 +1,7 @@
-import { AutomowerEvent, ErrorEvent } from '../../../src/clients/automower/automowerEventStreamClient';
-import { EventStreamServiceImpl } from '../../../src/services/husqvarna/eventStreamService';
+import { AutomowerEvent, ErrorEvent } from '../../../../src/clients/automower/automowerEventStreamClient';
+import { AutomowerEventStreamService } from '../../../../src/services/husqvarna/automower/automowerEventStreamService';
 
-export class EventStreamServiceImplSpy extends EventStreamServiceImpl {
+export class AutomowerEventStreamServiceSpy extends AutomowerEventStreamService {
     public shouldRunKeepAlive = true;
     public keepAliveExecuted = false;
